# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Tally's Barbershop & Beauty Studio API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Service {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  category: string;
  priceKes: number;
  durationMinutes: number;
  isActive: boolean;
  createdAt?: string;
}

export interface ServiceInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  category: string;
  priceKes: number;
  durationMinutes: number;
  isActive?: boolean;
}

export interface ServiceUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  category?: string;
  priceKes?: number;
  durationMinutes?: number;
  isActive?: boolean;
}

export interface Staff {
  id: number;
  name: string;
  role: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  specializations: string[];
  /** @nullable */
  rating?: number | null;
  commissionPct?: number;
  isActive: boolean;
  completedServices?: number;
  revenueGenerated?: number;
}

export interface StaffInput {
  /** @minLength 1 */
  name: string;
  role: string;
  bio?: string;
  photoUrl?: string;
  specializations?: string[];
  commissionPct?: number;
  isActive?: boolean;
}

export interface StaffUpdate {
  name?: string;
  role?: string;
  bio?: string;
  photoUrl?: string;
  specializations?: string[];
  commissionPct?: number;
  isActive?: boolean;
}

export interface Availability {
  staffId: number;
  date: string;
  slots: string[];
}

export interface Customer {
  id: number;
  name: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  gender: string | null;
  /** @nullable */
  birthday?: string | null;
  loyaltyTier: string;
  loyaltyPoints: number;
  totalVisits: number;
  totalSpentKes: number;
  createdAt?: string;
}

export interface CustomerInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  phone: string;
  email?: string;
  gender?: string;
  birthday?: string;
}

export interface CustomerUpdate {
  name?: string;
  phone?: string;
  email?: string;
  gender?: string;
  birthday?: string;
}

export interface Appointment {
  id: number;
  customerId: number;
  serviceId: number;
  staffId: number;
  date: string;
  timeSlot: string;
  status: string;
  /** @nullable */
  notes?: string | null;
  totalKes: number;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  serviceName?: string | null;
  /** @nullable */
  staffName?: string | null;
  createdAt?: string;
}

export interface AppointmentInput {
  customerId: number;
  serviceId: number;
  staffId: number;
  date: string;
  timeSlot: string;
  notes?: string;
}

export interface AppointmentUpdate {
  staffId?: number;
  date?: string;
  timeSlot?: string;
  status?: string;
  notes?: string;
}

export interface Review {
  id: number;
  customerId: number;
  /** @nullable */
  serviceId: number | null;
  /** @nullable */
  staffId?: number | null;
  rating: number;
  /** @nullable */
  comment?: string | null;
  status: string;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  serviceName?: string | null;
  createdAt?: string;
}

export interface ReviewInput {
  customerId: number;
  serviceId?: number;
  staffId?: number;
  rating: number;
  comment?: string;
}

export interface Loyalty {
  customerId: number;
  tier: string;
  points: number;
  totalVisits: number;
  /** @nullable */
  nextTier?: string | null;
  /** @nullable */
  visitsToNextTier?: number | null;
  benefits: string[];
}

export interface LoyaltyRedemption {
  points: number;
  reason?: string;
}

export interface DashboardSummary {
  dailyRevenue: number;
  weeklyRevenue: number;
  monthlyRevenue: number;
  totalCustomers: number;
  todayAppointments: number;
  activeAppointments: number;
  pendingAppointments: number;
  completedToday: number;
  repeatCustomerRate: number;
}

export interface RevenuePoint {
  date: string;
  revenue: number;
  appointments: number;
}

export interface ServicePopularity {
  serviceId: number;
  serviceName: string;
  category: string;
  bookings: number;
  revenueKes: number;
}

export interface StaffPerformance {
  staffId: number;
  name: string;
  completedServices: number;
  revenueKes: number;
  avgRating: number;
}

export type ListServicesParams = {
category?: string;
};

export type GetStaffAvailabilityParams = {
staffId: number;
date: string;
};

export type ListCustomersParams = {
search?: string;
};

export type ListAppointmentsParams = {
status?: string;
date?: string;
staffId?: number;
};

export type ListReviewsParams = {
serviceId?: number;
staffId?: number;
};