# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Tally's Barbershop & Beauty Studio API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Register a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not found
    patch:
      operationId: updateCustomer
      tags: [customers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}/appointments:
    get:
      operationId: getCustomerAppointments
      tags: [customers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"

  # ───────────────────── APPOINTMENTS ─────────────────────
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: date
          in: query
          required: false
          schema:
            type: string
        - name: staffId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Book an appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "404":
          description: Not found
    patch:
      operationId: updateAppointment
      tags: [appointments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    delete:
      operationId: cancelAppointment
      tags: [appointments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Cancelled
  /appointments/{id}/confirm:
    patch:
      operationId: confirmAppointment
      tags: [appointments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{id}/complete:
    patch:
      operationId: completeAppointment
      tags: [appointments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"

  # ───────────────────── REVIEWS ─────────────────────
...
  /loyalty/{customerId}:
    get:
      operationId: getLoyalty
      tags: [loyalty]
      summary: Get loyalty status for a customer
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loyalty"
  /loyalty/{customerId}/redeem:
    post:
      operationId: redeemLoyalty
      tags: [loyalty]
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoyaltyRedemption"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loyalty"

  # ───────────────────── DASHBOARD (wow endpoints) ─────────────────────
...
components:
  schemas:
...
    Customer:
      type: object
      required: [id, name, phone, gender, loyaltyTier, loyaltyPoints, totalVisits, totalSpentKes]
      properties:
        id:             { type: integer }
        name:           { type: string }
        phone:          { type: string }
        email:          { type: ["string", "null"] }
        gender:         { type: ["string", "null"] }
        birthday:       { type: ["string", "null"] }
        loyaltyTier:    { type: string }
        loyaltyPoints:  { type: integer }
        totalVisits:    { type: integer }
        totalSpentKes:  { type: integer }
        createdAt:      { type: string }
    CustomerInput:
      type: object
      required: [name, phone]
      properties:
        name:     { type: string, minLength: 1 }
        phone:    { type: string, minLength: 1 }
        email:    { type: string }
        gender:   { type: string }
        birthday: { type: string }
    CustomerUpdate:
      type: object
      properties:
        name:     { type: string }
        phone:    { type: string }
        email:    { type: string }
        gender:   { type: string }
        birthday: { type: string }

    # ─── APPOINTMENT ───
    Appointment:
      type: object
      required: [id, customerId, serviceId, staffId, date, timeSlot, status, totalKes]
      properties:
        id:           { type: integer }
        customerId:   { type: integer }
        serviceId:    { type: integer }
        staffId:      { type: integer }
        date:         { type: string }
        timeSlot:     { type: string }
        status:       { type: string }
        notes:        { type: ["string", "null"] }
        totalKes:     { type: integer }
        customerName: { type: ["string", "null"] }
        serviceName:  { type: ["string", "null"] }
        staffName:    { type: ["string", "null"] }
        createdAt:    { type: string }
    AppointmentInput:
      type: object
      required: [customerId, serviceId, staffId, date, timeSlot]
      properties:
        customerId: { type: integer }
        serviceId:  { type: integer }
        staffId:    { type: integer }
        date:       { type: string }
        timeSlot:   { type: string }
        notes:      { type: string }
    AppointmentUpdate:
      type: object
      properties:
        staffId:   { type: integer }
        date:      { type: string }
        timeSlot:  { type: string }
        status:    { type: string }
        notes:     { type: string }

    # ─── REVIEW ───
...
    Loyalty:
      type: object
      required: [customerId, tier, points, totalVisits, benefits]
      properties:
        customerId:  { type: integer }
        tier:        { type: string }
        points:      { type: integer }
        totalVisits: { type: integer }
        nextTier:    { type: ["string", "null"] }
        visitsToNextTier: { type: ["integer", "null"] }
        benefits:
          type: array
          items: { type: string }
    LoyaltyRedemption:
      type: object
      required: [points]
      properties:
        points: { type: integer }
        reason: { type: string }

    # ─── DASHBOARD ───
...