# lib/api-zod/src/generated/types/customerUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Tally's Barbershop & Beauty Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface CustomerUpdate {
  name?: string;
  phone?: string;
  email?: string;
  gender?: string;
  birthday?: string;
}