# lib/api-zod/src/generated/types/loyalty.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Tally's Barbershop & Beauty Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface Loyalty {
  customerId: number;
  tier: string;
  points: number;
  totalVisits: number;
  /** @nullable */
  nextTier?: string | null;
  /** @nullable */
  visitsToNextTier?: number | null;
  benefits: string[];
}